Move scripts to insight_face folder
