+++ insight_face/config.py
